chatapp/requirements.txt
textual==0.48.2
httpx==0.26.0
websockets==12.0
cryptography==42.0.0
python-dotenv==1.0.0
